01-llm-api/pyproject.toml
[project]
name = "llm-cli-bot"
version = "0.1.0"
description = "Educational CLI bot for LLM interaction with rich formatting and metrics tracking"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = []
